requirements.txt
flask==3.0.3
flask-cors==5.0.0
pillow==11.0.0
numpy==2.1.3
tensorflow==2.20.0
